+++ src/internal/checksum.rs
@@ -1,5 +1,3 @@
-// ========================================================================= //
-
 pub struct Checksum {
     value: u32,
     remainder: u32,
@@ -44,8 +42,6 @@ impl Checksum {
     }
 }
 
-// ========================================================================= //
-
 #[cfg(test)]
 mod tests {
     use super::Checksum;
@@ -81,5 +77,3 @@ mod tests {
         assert_eq!(checksum.value(), 0x30a65abd);
     }
 }
-
-// ========================================================================= //
